Enables Azure Key Vault integration
Configures the application to retrieve secrets from Azure Key Vault.
This allows for secure storage and management of sensitive information,
such as API keys and connection strings.

Adds support for Managed Identity authentication when accessing Key Vault,
allowing the application to authenticate without hardcoded credentials.

Removes appsettings.Development.json from .gitignore and adds it back to the project.
This file contains the sensitive information, and will be kept in KeyVault.